+++ src/views/Login.vue
@@ -1,14 +1,16 @@
 <template>
     <div class="flex flex-col grow h-full w-full justify-evenly items-center">
         <div class="flex flex-col justify-center items-center">
-            <form-card
+            <!-- LOGIN CARD -->
+            <FormCard
+                v-if="mode === 'login'"
                 class="show-up p-2"
-                :title="mode === 'login' ? Lang.CreateTranslationContext('verbs', 'LogIn'): Lang.CreateTranslationContext('verifications', 'passwordReset')"
-                :validate="mode === 'login' ? Lang.CreateTranslationContext('verbs', 'LogIn'): Lang.CreateTranslationContext('verbs', 'Send')"
+                :title="Lang.CreateTranslationContext('verbs', 'LogIn')"
+                :validate="Lang.CreateTranslationContext('verbs', 'LogIn')"
                 :on-cancel="cancel"
-                :on-validate="mode === 'login' ? login: sendMail"
+                :on-validate="login"
             >
-                <input-text
+                <InputText
                     ref="email"
                     name="email"
                     :label="Lang.CreateTranslationContext('account', 'Email')"
@@ -16,7 +18,7 @@
                     class="show-down"
                     style="animation-delay: 100ms;"
                 />
-                <input-text
+                <InputText
                     ref="password"
                     class="show-down"
                     :class="mode === 'login' ? '': 'hidden'"
@@ -28,13 +30,31 @@
                 />
                 <button
                     class="show-down italic text-slate-500 dark:text-slate-400 hover:underline"
-                    :class="mode === 'login' ? '': 'hidden'"
                     style="animation-delay: 300ms;"
                     :onclick="toggleMode"
                 >
-                    <get-text :context="Lang.CreateTranslationContext('account', 'ForgotPassword')" />
+                    <GetText :context="Lang.CreateTranslationContext('account', 'ForgotPassword')" />
                 </button>
-            </form-card>
+            </FormCard>
+
+            <!-- FORGOT PASSWORD CARD -->
+            <FormCard
+                v-else
+                class="show-up p-2"
+                :title="Lang.CreateTranslationContext('verifications', 'passwordReset')"
+                :validate="Lang.CreateTranslationContext('verbs', 'Send')"
+                :on-cancel="cancel"
+                :on-validate="sendForgotPasswordMail"
+            >
+                <InputText
+                    ref="email"
+                    name="email"
+                    :label="Lang.CreateTranslationContext('account', 'Email')"
+                    :placeholder="Lang.CreateTranslationContext('account', 'Email')"
+                    class="show-down"
+                    style="animation-delay: 100ms;"
+                />
+            </FormCard>
         </div>
     </div>
 </template>
@@ -43,7 +63,8 @@
 import FormCard from '../components/cards/FormCard.vue';
 import InputText from '../components/inputs/InputText.vue';
 import GetText from '../components/text/GetText.vue';
-import API from '../scripts/API';
+import { API } from '../scripts/API';
+import ROUTES from '../scripts/routes';
 import { redirectLink } from '../scripts/common';
 import Lang from '../scripts/Lang';
 import { Log } from '../scripts/Logs';
@@ -72,33 +93,6 @@ export default {
             if (this.mode === 'login') this.$router.go(-1);
             else this.mode = 'login';
         },
-        async sendMail(form) {
-            const log = form.log(await Lang.GetTextAsync(Lang.CreateTranslationContext('verbs', 'Sending')), Log.INFO);
-
-            const forgotEmail = form.body().email;
-            if (!forgotEmail) {
-                log.update(await Lang.GetTextAsync(Lang.CreateTranslationContext('account', 'SpecifyEmail')), Log.WARNING);
-                form.focus('email');
-                setTimeout(() => { log.delete(); }, 4000);
-                return;
-            }
-
-            try {
-                const res = await API.execute(
-                    API.ROUTE.RESET.PASSWORD() + '?email=' + forgotEmail,
-                    API.METHOD.GET
-                );
-
-                log.update(res.message, Log.SUCCESS);
-                redirectLink(this, true);
-                setTimeout(() => { log.delete(); }, 2000);
-            } catch (err) {
-                log.error(err);
-                if (err.field) form.focus(err.field);
-                console.error(err);
-                setTimeout(() => { log.delete(); }, 4000);
-            }
-        },
         async login(form) {
             const log = form.log(await Lang.GetTextAsync(Lang.CreateTranslationContext('verbs', 'LoggingIn')), Log.INFO);
 
@@ -117,12 +111,19 @@ export default {
             }
 
             try {
-                const resLogin = await API.execute(API.ROUTE.LOGIN(), API.METHOD.POST, { email: body.email, password: body.password });
+                const resLogin = await API.Request(ROUTES.AUTH.LOGIN(body.email, body.password));
+                if (resLogin.error) {
+                    log.error(resLogin);
+                    if (resLogin.field) form.focus(resLogin.field);
+                    setTimeout(() => { log.delete(); }, 4000);
+                    return;
+                }
+
                 const user = new User();
                 user.setTokens(resLogin.data);
                 user.save();
 
-                const resMe = await API.execute_logged(API.ROUTE.ME());
+                const resMe = await API.RequestLogged(ROUTES.USERS.ME.GET());
                 user.setInformations(resMe.data);
                 user.save();
 
@@ -130,14 +131,35 @@ export default {
                 redirectLink(this, true);
                 setTimeout(() => { log.delete(); }, 2000);
             } catch (err) {
-                log.error(err);
-                if (err.field) form.focus(err.field);
                 console.error(err);
-                setTimeout(() => { log.delete(); }, 4000);
             }
         },
-        forgotPassword() {
-            
+        async sendForgotPasswordMail(form) {
+            const log = form.log(await Lang.GetTextAsync(Lang.CreateTranslationContext('verbs', 'Sending')), Log.INFO);
+
+            const forgotEmail = form.body().email;
+            if (!forgotEmail) {
+                log.update(await Lang.GetTextAsync(Lang.CreateTranslationContext('account', 'SpecifyEmail')), Log.WARNING);
+                form.focus('email');
+                setTimeout(() => { log.delete(); }, 4000);
+                return;
+            }
+
+            try {
+                const res = await API.Request(ROUTES.PASSWORD.SENDRESETMAIL(forgotEmail));
+                if (res.error) {
+                    log.error(res);
+                    if (res.field) form.focus(res.field);
+                    setTimeout(() => { log.delete(); }, 4000);
+                    return;
+                }
+
+                log.update(res.message, Log.SUCCESS);
+                redirectLink(this, true);
+                setTimeout(() => { log.delete(); }, 2000);
+            } catch (err) {
+                console.error(err);
+            }
         }
     }
 }
